Weitere Fragen funktionieren (Stanford Parser) Es wird nun ebenfalls nach VBN und VBP gesucht Ansonsten auch noch nach VB[[:space:]] Es wird nun beim Stanford Parser überprüft, ob ein Subjekt in der ersten Zeile steht
Das Fragezeichen am Ende bei kurzen Fragen wird nicht mehr mitgenommen
Das Dollarzeichen bei Possesivpronomen wird nicht mehr in result.txt aufgenommen
Wenn keine Zeilen mehr da sind, beendet die Objektsuche ihre Arbeit

ReVerb ist nun richtig eingestellt und hat auch mehr von den Testfragen verarbeiten können

+++ process_question_stanford.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-debug=0
+debug=1
 
 
 if [ $# -ne 2 ]; then
@@ -30,24 +30,54 @@ if [[ -z "$(cat $resultfile)" ]]; then
 	echo "No Open IE Extraction"
 	#Keine Extraktionen ...
 
-	lexparser.sh "$qfile" 2> /dev/null > result.txt
+	#Dollarzeichen der Annotationen von Possesivprononmen entfernen mit sed
+	lexparser.sh "$qfile" 2> /dev/null | sed -E 's/\$//g' > result.txt
 
-	grep -A 1 -B 1 -e VBD -e VBZ result.txt > tmp_relation.txt
+	# head: bei mehreren Funden nur den ersten betrachten
+	grep -A 1 -B 1 -e VBD -e VBZ -e VBN -e VBP result.txt | head -n 3 > tmp_relation.txt
 
-	object_line=$(awk 'NR==3' tmp_relation.txt)
-	if [[ "$object_line" != *NN* ]]; then
-		grep -A 10 -B 1 -e VBD -e VBZ result.txt > tmp_relation.txt
-		awk 'NR>2{if(index($0, "NN") > 0){printf $0" ";exit;}else{printf $0}}' tmp_relation.txt > tmp_relation2.txt
-		./clean_parsing.sh nogrep tmp_relation2.txt | sed -E 's/ +/ /g' > tmp_relation3.txt
-		./clean_parsing.sh nogrep tmp_relation.txt | head -n 2 > $resultfile
-		cat tmp_relation3.txt >> $resultfile
+	if [[ ! -z "$(cat tmp_relation.txt)" ]]; then
+		object_line=$(awk 'NR==3' tmp_relation.txt)
+		# Steht in der dritten Zeile kein Substantiv?
+		if [[ "$object_line" != *NN* ]]; then
+			# Wenn nicht, sollen die Zeilen danach konkateniert werden bis eines da ist (Brechstange)
+			grep -A 10 -B 1 -e VBD -e VBZ -e VBN -e VBP result.txt > tmp_relation.txt
+			awk 'NR>2{if(index($0, "NN") > 0){printf $0" ";exit;}else{printf $0}} !NF{exit}' tmp_relation.txt > tmp_relation2.txt
+			# der xargs Aufruf ist hier eingefügt, um leading und trailing whitespace zu entfernen
+			./clean_parsing.sh nogrep tmp_relation2.txt | sed -E 's/ +/ /g' | xargs > tmp_relation3.txt
+			./clean_parsing.sh nogrep tmp_relation.txt | head -n 2 > $resultfile
+			cat tmp_relation3.txt >> $resultfile
 
-		rm tmp_relation2.txt
-		rm tmp_relation3.txt
+			rm tmp_relation2.txt
+			rm tmp_relation3.txt
+		else
+			./clean_parsing.sh nogrep tmp_relation.txt > $resultfile
+		fi
+
+		subject_line=$(awk 'NR==1' tmp_relation.txt)
+		# Steht in der ersten Zeile kein Substantiv?
+		if [[ "$subject_line" != *NN* ]]; then
+			# Wenn nicht, soll die Zeile davor konkateniert werden (Brechstange)
+			# im Gegensatz zum Vorgehen beim Objekt weiter oben wird hier nicht überprüft, ob die neue
+			# Zeile ein Substantiv enthält - wenn es noch eine Zeile weiter oben ist, ist das eben so
+			grep -A 1 -B 2 -e VBD -e VBZ -e VBN -e VBP result.txt > tmp_relation.txt
+			# Ist denn jetzt ein Substantiv in der ersten Zeile?
+			if [[ "$(head -n 1 tmp_relation.txt)" == *NN* ]]; then
+				awk 'NR!=1{printf $0" ";exit;} NR==1{printf $0}' tmp_relation.txt > tmp_relation2.txt
+				./clean_parsing.sh nogrep tmp_relation2.txt | sed -E 's/ +/ /g' > tmp_relation3.txt
+				# tmp_relation2.txt können wir hier überschreiben
+				mv $resultfile tmp_relation2.txt
+				cat tmp_relation2.txt | tail -n 2 >> tmp_relation3.txt
+				mv tmp_relation3.txt $resultfile
+
+				rm tmp_relation2.txt
+			fi
+		fi
+	else
+		./clean_parsing.sh 1 "-e VB[[:space:]]" > $resultfile
 	fi
 
-	rm tmp_relation.txt
-	rm result.txt
+	
 
 #	if [[ -z "$(cat tmp_relation.txt)" ]]; then
 #		# nach VB suchen
@@ -65,5 +95,8 @@ if [[ -z "$(cat $resultfile)" ]]; then
 	if [[ -z "$(cat $resultfile)" ]]; then
 		echo "Aaaahh"
 		exit 1
+	elif [ $debug -eq 0 ]; then
+		rm tmp_relation.txt
+		rm result.txt
 	fi
 fi
